requirements/base.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
Django==5.0.3
django-4-jet==1.0.9
django-autocomplete-light==3.11.0
django-better-admin-arrayfield==1.4.2
django-cors-headers==4.3.1
django-debug-toolbar==6.1.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.1
django-js-asset==2.2.0
django-mptt==0.16.0
django-querysetsequence==0.17
django-rest-framework-social-oauth2==1.2.0
djangorestframework==3.15.0
djangorestframework-jsonapi==6.1.0
idna==3.11
inflection==0.5.1
jwcrypto==1.5.6
oauthlib==3.3.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.3
social-auth-core==4.8.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0

# autenticación
django-oauth2-provider==0.2.6.1
django-oauth-toolkit==2.4.0
mozilla-django-oidc==4.0.0
